Extract shared Markdown list item formatter
