+++ src/lib/hig/render.ts
@@ -2,6 +2,7 @@
  * Human Interface Guidelines (HIG) rendering functionality
  */
 
+import { formatListItem } from "../markdown/list"
 import { extractTitleFromIdentifier } from "../reference/render"
 import type { ContentItem, TextFragment } from "../types"
 import type {
@@ -486,30 +487,6 @@ function mapHIGAsideStyleToCallout(style: string): string {
     default:
       return "NOTE"
   }
-}
-
-/**
- * Format rendered content as a Markdown list item, indenting continuation
- * lines (including nested lists) under the marker.
- */
-function formatListItem(marker: string, content: string): string {
-  const trimmed = content.replace(/\n+$/, "")
-  if (!trimmed) {
-    return `${marker}\n`
-  }
-
-  // Only collapse blank lines immediately before a nested list; keep other
-  // paragraph breaks. Indent every continuation line (including blanks) so
-  // nested blocks stay inside the list item.
-  const tightened = trimmed.replace(/\n{2,}(?=[-*] |\d+\. )/g, "\n")
-  const lines = tightened.split("\n")
-  const indent = " ".repeat(marker.length)
-  const [first, ...rest] = lines
-  let result = `${marker}${first}`
-  for (const line of rest) {
-    result += `\n${indent}${line}`
-  }
-  return `${result}\n`
 }
 
 /**
